chore: separate some logic
